Fix inline fragment syntax highlighting
The "on" keyword was previously part of the graphqlStructure match
group, but that group is excluded from being used within graphqlFold
sections because "structure" here refers to "outer" (not "inline")
structure.

This change moves "on" to a new "graphqlInlineStructure" match group
that also highlights as "Structure" but that is allowed to appear within
folds.

This addresses part of the issue reported in #5.