Generalize 'on' using a new graphqlKeyword group
In 1d21bed, I had forgotten that 'on' is also used to declare a type
condition. It was highlighted correctly but was in the semantically-
incorrect `graphqlInlineStructure` highlight group.

This new approach introduces a new generalized `graphqlKeyword` group
that correctly matches 'on' as the type condition keyword in all
contexts.