Use an autoloaded function to get JavaScript tags
We previously relied on the main plugin script to set the global
`g:graphql_javascript_tags` variable if it wasn't already set by the
user. However, there are some legitimate cases when the plugin script
wouldn't be loaded before the `after/` syntax files (see #25), resulting
in the global variable not being set at all.

Using an autoload script addresses that shortcoming because vim will
load it on-demand when the `after/` syntax files are sourced.