Remove the graphqlConstant syntax definition
Constants don't exist in the GraphQL language. I originally added this
syntax definition to cover `enum` values (which are written in ALL_CAPS
by convention) and directive locations.

Unfortunately, this syntax unintentionally matched a few other cases,
such as standard object names that happened to be spelled with CAPS.

Because it's not an official part of the GraphQL grammar, I decided the
best approach was to remove graphqlConstant. In 4c2fdadc, I added
explicit support for directive locations, so that case is now handled
correctly. `enum` values will now just be highlighted as graphqlType,
which is reasonable.