Fix reason syntax (#60)
* Fix reason syntax

The change was incorrect and due to the change, the highlighting didn't work anymore. We cannot match within a multiline string because it needs to match an "extension point" everything that happens inside that extension point is basically taken care of by the language extension. I also removed the extra items in contains, I am not sure why that needs to be there because it contains no reason at all, just GraphQL syntax. Also, the current change is independent of the reason-syntax you use.

* A few tweaks to make the tests pass

- Be more lenient with whitespace matches (more still to do)
- Remove the vim-reasonml-specific syntax check
- Remove the unnecessary 'filetype' setting

Co-authored-by: Jon Parise <jon@indelible.org>