Attempt to honor indentation within templates
When we're within a template string, it's common for the author to
indent the GraphQL syntax some level beyond the outer scope.

Attempt to honor that initial indentation by looking for the first line
of GraphQL syntax and telling vim to leave its indentation untouched
(-1). The following lines will be indentation relative to this initial
line.