Recognize GraphQL strings passed to JavaScript functions
Some JavaScript GraphQL frameworks use well-known functions rather than
tagged template literals. This change introduces basic support for these
cases via a new `g:graphql_javascript_functions` variable (["graphql"]).

This is implemented as a bit of a hack on top of the existing tagged
template support to avoid introducing new syntax regions:

    Tagged template:    graphql`
    Function name:      graphql(`
    Syntax pattern:     (graphql|graphql\()`

This is a good-enough place to start, and the overall approach gives us
room to iterate later if we want to revisit the implementation.