Reorganize JavaScript syntax groups (#113)
This is a significant reorganization of the JavaScript-related syntax
groups that's intended to make this (hairy) code a bit more predictable
and easier to maintain.

First, we now always define the syntax groups based on vim's runtime
support for JavaScript. These syntax definitions are always available
in vim 8.2+, which is a reasonable baseline these days.

Second, when the optional vim-javascript plugin is detected, we add
syntax support for its groups and then relink the shared highlights to
its syntax types. This maintains support for vim's JavaScript groups
while also supporting vim-javascript and using its enhanced highlights.

Lastly, while I'm already shuffling so many lines around, wrap them for
readability.